Added Input Handler for A-Z, STR, CTRL, ALT, TAB

+++ Droid/src/util/Point.java
@@ -12,7 +12,7 @@ public class Point {
 		this.y = 0;
 	}
 	
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
